replacing fetch_array with fetch_assoc
fetch_array returns an array that contains the results of fetch_row and fetch_assoc making it twice as big than need be.
I don't think we have a case where we use both at the same time, so we shouldn't use fetch_array. We should use fetch_assoc and where we can fetch_row.